requirements.txt
fastapi
uvicorn
requests
apscheduler
motor
pydantic[email]
python-dotenv
